fix: :card_file_box: Changed FallingTimer Component to Resource
Issue: #21
