+++ src/block/movement.rs
@@ -5,10 +5,18 @@ use crate::player::{
     BlockDirection,
     BlockMoveEvent,
 };
-use super::{
-    PlayerBlock,
-    FallingTimer,
-};
+use super::PlayerBlock;
+
+const FPS: f32 = 0.2;
+
+#[derive(Resource, Deref, DerefMut)]
+struct FallingTimer(Timer);
+
+impl Default for FallingTimer {
+    fn default() -> Self {
+        Self(Timer::from_seconds(FPS, TimerMode::Repeating))
+    }
+}
 
 fn movement(
     mut events: EventReader<BlockMoveEvent>,
@@ -27,14 +35,15 @@ fn movement(
 }
 
 fn falling(
-    mut query: Query<(&mut FallingTimer, &mut Transform), With<PlayerBlock>>,
+    mut timer: ResMut<FallingTimer>,
+    mut query: Query<&mut Transform, With<PlayerBlock>>,
     time: Res<Time>,
 ) {
-    for (mut timer, mut transform) in &mut query {
-        timer.tick(time.delta());
-        if !timer.just_finished() { continue }
-        timer.reset();
-        // debug!("movement");
+    timer.tick(time.delta());
+    if !timer.just_finished() { return }
+    timer.reset();
+    // debug!("falling block");
+    for mut transform in &mut query {
         transform.translation.y -= GRID_SIZE;
     }
 }
@@ -44,6 +53,7 @@ pub struct MovementPlugin;
 impl Plugin for MovementPlugin {
     fn build(&self, app: &mut App) {
         app
+            .insert_resource(FallingTimer::default())
             .add_systems(Update, (
                 movement,
                 falling,
